requirements.txt
black==22.3.0
click==8.1.3
dominate==2.6.0
Flask==2.1.2
Flask-Bootstrap4==4.0.2
Flask-Cors==3.0.10
importlib-metadata==4.11.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mypy-extensions==0.4.3
pathspec==0.9.0
pkg_resources==0.0.0
platformdirs==2.5.2
pycryptodome==3.14.1
six==1.16.0
tomli==2.0.1
typing_extensions==4.2.0
visitor==0.1.3
Werkzeug==2.1.2
zipp==3.8.0
